Rename sigma_cx_h0 to sigma_cx_h0.py
In the original file, a variable 'ty' was set to type_of(E,nDim=nDim) but never used. I think all that line did was set nDim, since that was referenced later in the code. Since we're currently using type() instead of type_of(), I replaced this with the is_scalar variable - this can be changed later if we decide to fully translate type_of()