Update KN1D.py
A few updates, including the next section of code. Some of what was in this section had to be skipped until we finish other functions (kinetic_h, kinetic_h2), or because we're still figuring out how to implement some features (saving/reading files). Also made a few edits to earlier portions which are explained in the code, and also added to the import lines.